moving sync var hooks tests to their own class
