+++ Assets/Mirror/Tests/Editor/Weaver/WeaverSyncVarHookTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8b8a829462efd04479519a188e458616
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
